+++ src/agent_c_api_ui/agent_c_react_client/src/components/chat_interface/ChatInterface.jsx
@@ -672,27 +672,27 @@ const ChatInterface = ({sessionId, customPrompt, modelName, modelParameters, onP
 
     return (
         <Card className="flex flex-col h-full bg-white/50 dark:bg-gray-800/50 backdrop-blur-sm border dark:border-gray-700 shadow-lg rounded-xl relative z-0">
-            <ScrollArea className="flex-1 px-4 py-3 min-h-[400px]">
-                {/* Add copy entire chat button */}
-                <div className="flex justify-end gap-2 p-2 sticky top-0 z-10 bg-white/80 dark:bg-gray-800/80 backdrop-blur-sm">
-                    <CopyButton
-                        content={getChatCopyContent}
-                        htmlContent={getChatCopyHTML}
-                        tooltipText="Copy entire chat"
-                        successText="Chat copied!"
-                        size="sm"
-                        variant="outline"
-                        className="border-gray-300"
-                    />
-                    <ExportHTMLButton
-                        messages={messages}
-                        tooltipText="Export as HTML"
-                        filename={`chat-export-${new Date().toISOString().slice(0, 10)}.html`}
-                        size="sm"
-                        variant="outline"
-                        className="border-gray-300"
-                    />
-                </div>
+            {/* Add copy entire chat button - fixed at top */}
+            <div className="flex justify-end gap-2 p-2 sticky top-0 z-10 bg-white/80 dark:bg-gray-800/80 backdrop-blur-sm border-b border-gray-100 dark:border-gray-700">
+                <CopyButton
+                    content={getChatCopyContent}
+                    htmlContent={getChatCopyHTML}
+                    tooltipText="Copy entire chat"
+                    successText="Chat copied!"
+                    size="sm"
+                    variant="outline"
+                    className="border-gray-300"
+                />
+                <ExportHTMLButton
+                    messages={messages}
+                    tooltipText="Export as HTML"
+                    filename={`chat-export-${new Date().toISOString().slice(0, 10)}.html`}
+                    size="sm"
+                    variant="outline"
+                    className="border-gray-300"
+                />
+            </div>
+            <ScrollArea className="flex-1 px-4 py-3">
 
                 <div className="space-y-4">
                     {messages.map((msg, idx) => {
